fix types

+++ src/Open3D/Core/Hashmap/Hashmap.h
@@ -48,14 +48,14 @@
 namespace open3d {
 
 template <typename Hash = DefaultHash, typename KeyEq = DefaultKeyEq>
-std::shared_ptr<Hashmap<Hash, KeyEq>> CreateHashmap(uint32_t init_buckets,
-                                                    uint32_t dsize_key,
-                                                    uint32_t dsize_value,
+std::shared_ptr<Hashmap<Hash, KeyEq>> CreateHashmap(size_t init_buckets,
+                                                    size_t dsize_key,
+                                                    size_t dsize_value,
                                                     open3d::Device device) {
     static std::unordered_map<
             Device::DeviceType,
             std::function<std::shared_ptr<Hashmap<Hash, KeyEq>>(
-                    uint32_t, uint32_t, uint32_t, Device)>,
+                    size_t, size_t, size_t, Device)>,
             utility::hash_enum_class::hash>
             map_device_type_to_hashmap_constructor = {
                 {Device::DeviceType::CPU, CreateCPUHashmap<Hash, KeyEq>},
